requirements.txt
fastapi==0.95.0
pydantic[email]==1.10.7
uvicorn[standard]==0.21.1
python-jose==3.3.0
sqlalchemy==2.0.7
passlib==1.7.4
python-dotenv==1.0.0
bcrypt==4.0.1
emails==0.6
python-multipart==0.0.6
pytest==7.2.2
httpx==0.23.3
tenacity==8.2.2
psycopg2-binary==2.9.6